+++ ADictionaryEcho/app/src/main/java/ayds/dictionary/echo/model/business/ExceptionHandlerImpl.java
@@ -3,7 +3,7 @@ package ayds.dictionary.echo.model.business;
 import java.util.Map;
 
 import ayds.dictionary.echo.model.TranslatorModelExceptionListener;
-import ayds.dictionary.echo.model.business.services.Source;
+import ayds.dictionary.echo.model.services.Source;
 import ayds.dictionary.echo.model.exceptions.ModelNoConnectionException;
 import ayds.dictionary.echo.model.exceptions.NonTranslatableWordException;
