+++ ADictionaryEcho/app/src/main/java/ayds/dictionary/echo/model/business/NullTranslationConcept.java
@@ -1,6 +1,6 @@
 package ayds.dictionary.echo.model.business;
 
-import ayds.dictionary.echo.model.business.services.Source;
+import ayds.dictionary.echo.model.services.Source;
 
 public class NullTranslationConcept extends TranslationConcept {
