+++ ADictionaryEcho/app/src/main/java/ayds/dictionary/echo/model/services/ServiceAdministrator.java
@@ -1,0 +1,8 @@
+package ayds.dictionary.echo.model.services;
+
+import java.util.Set;
+
+public interface ServiceAdministrator {
+    Set<Source> getServices();
+    String getMeaningBySource(Source source,String wordToTranslate) throws Exception;
+}
