+++ ADictionaryEcho/app/src/main/java/ayds/dictionary/echo/model/services/ServiceDefinition.java
@@ -1,4 +1,4 @@
-package ayds.dictionary.echo.model.business.services;
+package ayds.dictionary.echo.model.services;
 
 public interface ServiceDefinition {
     String getResult(String wordToGetResult) throws Exception;
