+++ ADictionaryEcho/app/src/main/java/ayds/dictionary/echo/model/services/Source.java
@@ -1,4 +1,4 @@
-package ayds.dictionary.echo.model.business.services;
+package ayds.dictionary.echo.model.services;
 
 public enum Source {
     NULL(""),
